Refactor project to have api clients separated into classes. Updated docs, examples, and README. Switched project to pytest and pytest-runner, see README for info
