+++ docs/api_docs.rst
@@ -8,3 +8,29 @@ gbdxtools.interface module
     :members:
     :undoc-members:
     :show-inheritance:
+
+.. automodule:: gbdxtools.catalog
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: gbdxtools.s3
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: gbdxtools.ordering
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: gbdxtools.workflow
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: gbdxtools.idaho
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
