--- integration_test.py
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-
-import tests
-
-tests.run_tests()
