+++ setup.py
@@ -26,5 +26,7 @@ setup(name='gbdxtools',
                         'boto==2.39.0',
                         'gbdx-auth==0.1.2',
                         'Pillow==3.1.1',
-                        'pygeoif==0.6']
+                        'pygeoif==0.6'],
+      setup_requires=['pytest-runner'],
+      tests_require=['pytest']
       )
