+++ tests/unit/test_interface.py
@@ -1,0 +1,6 @@
+from gbdxtools import Interface
+
+def test_init():
+    gi = Interface()
+    assert isinstance(gi, Interface)
+
